Complete fix for 3 critical Jupyter protocol bugs and build optimization.

This merge brings:
- LogEnv configuration fix (respects --log-level and HSJUPYTER_LOG_LEVEL)
- Protocol delimiter parsing fix (accepts empty frame per Jupyter spec)
- Reply envelope construction fix (uses signaturePayloadFrom)
- Build performance optimization with ld.gold (33x speedup: 77s → 2.3s)
- Complete CHANGELOG.md with project history
- Updated documentation (code comments, build-performance.md, AGENTS.md)

Test Results:
- 248 unit tests passing
- 60 integration tests passing
- kernel_info_request/reply validated

Spec: specs/005-fix-jupyter-kernel-bug/
Tasks: 19/19 complete

# Conflicts:
#	src/HsJupyter/Router/RequestRouter.hs
#	test_kernel.py
#	validation_test.hs

+++ src/HsJupyter/Runtime/ResourceGuard.hs
@@ -204,9 +204,9 @@ monitorMemoryUsage action = do
           usedBytes = max 0 (endBytes - startBytes)
           bytesPerMb :: Double
           bytesPerMb = 1024 * 1024
-          rawMb :: Integer
+          rawMb :: Int
           rawMb = ceiling (fromIntegral usedBytes / bytesPerMb)
-          usedMB = if rawMb <= 0 then 1 else fromIntegral rawMb
+          usedMB = if rawMb <= 0 then 1 else rawMb
 
       pure $ case result of
         Left ex    -> Left ex
